testing and debugging

+++ src/NumpyDotNet/NumpyDotNet/Extensions.cs
@@ -352,11 +352,10 @@ namespace NumpyDotNet
             return np.std(a, axis, dtype);
         }
 
-        public static ndarray partition(this ndarray a, IEnumerable<int> kth, int? axis = null, string kind = "introselect", IEnumerable<string> order = null)
+        public static ndarray partition(this ndarray a, IEnumerable<long> kth, int? axis = null, string kind = "introselect", IEnumerable<string> order = null)
         {
             return np.partition(a, kth, axis, kind, order);
         }
-
 
         public static List<T> ToList<T>(this ndarray a)
         {
